backend/requirements.txt
asgiref==3.8.1
Django==5.0.14
django-auth-ldap==5.2.0
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
inflection==0.5.1
mssql-django==1.5
numpy==2.3.2
packaging==25.0
pandas==2.3.1
pillow==11.2.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyJWT==2.9.0
pyodbc==5.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-ldap==3.4.4
pytz==2025.2
PyYAML==6.0.2
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
